🐛 fix: address PR feedback - ref-counted URLs and optimized persistence
