+++ apps/web/src/lib/components/GraphView.svelte
@@ -285,7 +285,7 @@
     }
     if (imageManager) {
       imageManager.destroy({
-        releaseImageUrl: (path) => vault.releaseImageUrl(path),
+        releaseImageUrl: (path: string) => vault.releaseImageUrl(path),
       } as any);
       imageManager = undefined;
     }
@@ -308,7 +308,7 @@
         didFinalizeLoad = false;
         if (imageManager)
           imageManager.destroy({
-            releaseImageUrl: (path) => vault.releaseImageUrl(path),
+            releaseImageUrl: (path: string) => vault.releaseImageUrl(path),
           } as any);
       });
     }
@@ -408,7 +408,7 @@
         imageManager!.sync({
           showImages: graph.showImages,
           resolveImageUrl: (path) => vault.resolveImageUrl(path),
-          releaseImageUrl: (path) => vault.releaseImageUrl(path),
+          releaseImageUrl: (path: string) => vault.releaseImageUrl(path),
           onBatchApplied: (count) => {
             debugStore.log(
               `[GraphView] Applied ${count} images to graph nodes.`,
